Fixed creating multiple connections when making new connection and on garden sync.
